+++ base/deprecated.jl
@@ -1350,6 +1350,9 @@ end
 
 @deprecate which(s::Symbol) which(Main, s)
 
+# PR #26013
+@deprecate_binding DenseArray AbstractStridedArray
+
 @deprecate IOBuffer(data::AbstractVector{UInt8}, read::Bool, write::Bool=false, maxsize::Integer=typemax(Int)) IOBuffer(data, read=read, write=write, maxsize=maxsize)
 @deprecate IOBuffer(read::Bool, write::Bool) IOBuffer(read=read, write=write)
 @deprecate IOBuffer(maxsize::Integer) IOBuffer(read=true, write=true, maxsize=maxsize)